SDIT-3215: ⬆️ More upgrading to spring boot 4

+++ hmpps-prisoner-search-indexer/src/main/kotlin/uk/gov/justice/digital/hmpps/prisonersearch/indexer/repository/PrisonerRepository.kt
@@ -147,8 +147,8 @@ class PrisonerRepository(
     val response = openSearchRestTemplate.update(
       UpdateQuery.builder(prisonerNumber)
         .withDocument(Document.from(prisonerMap))
-        .withIfSeqNo(summary.sequenceNumber.toInt())
-        .withIfPrimaryTerm(summary.primaryTerm.toInt())
+        .withIfSeqNo(summary.sequenceNumber)
+        .withIfPrimaryTerm(summary.primaryTerm)
         .build(),
       IndexCoordinates.of(PRISONER_INDEX),
     )
